+++ CHANGELOG.md
@@ -14,9 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 *  High fidelity thermal conductivity models for H2 and N2 [#72](@ref)
 *  `SetFields` utility that allows to set a field to desired value within a box / circle / sphere [#73](@ref)
 *  Helmholtz Energy equation of state and supporting framework for it for H2 and N2 [#75](@ref)
-*  Time-average and RMS of Vector and Scalar fields [#78]
+*  `FieldAverage` and `FieldRMS` of Vector and Scalar fields [#78]
 *  Added `RotatingWall` velocity boundary condition [#81](@ref)
 *  Added `KOmegaSST` turbulence model [#82](@ref)
+*  Added capablity to write out TensorFields to .vtk/vtu including Reynolds Stress Tensor [#87](@ref)
 
 
 ### Fixed